+++ jexxa-core/src/test/java/io/jexxa/TestConstants.java
@@ -4,4 +4,9 @@ public class TestConstants
 {
     public static final String UNIT_TEST = "unit-test";
     public static final String INTEGRATION_TEST = "integration-test";
+
+    public static final String JEXXA_APPLICATION_SERVICE =  "io.jexxa.application.applicationservice";
+    public static final String JEXXA_DOMAIN_SERVICE =  "io.jexxa.application.domainservice";
+    public static final String JEXXA_DRIVEN_ADAPTER = "io.jexxa.application.infrastructure.drivenadapter";
+    public static final String JEXXA_DRIVING_ADAPTER = "io.jexxa.infrastructure.drivingadapter";
 }